Enhance error handling and validation in database modules
- Added a ValueError raise for missing step_range in filter_scope when scope is STEP_RANGE.
- Improved error handling in SessionManager to raise a SQLAlchemyError with a clear message if last_error is None.
- Refactored simulation comparison functions to streamline output formatting and ensure consistent return types.
- Updated test cases to reflect changes in error handling and validation logic across various database modules.

+++ farm/database/analyzers/analysis_utils.py
@@ -190,7 +190,7 @@ def calculate_consistency(frequencies: np.ndarray) -> float:
     """
     if len(frequencies) < 2:
         return 1.0
-    return 1.0 - np.std(frequencies) / (np.mean(frequencies) + 1e-10)
+    return float(1.0 - np.std(frequencies) / (np.mean(frequencies) + 1e-10))
 
 
 def calculate_periodicity(data: np.ndarray) -> float:
